Split tests for editableform and overview
