+++ packages/venia-concept/src/components/Checkout/__tests__/overview.spec.js
@@ -1,0 +1,106 @@
+import React from 'react';
+import { act } from 'react-test-renderer';
+import { createTestInstance } from '@magento/peregrine';
+
+import Overview from '../overview';
+import Section from '../section';
+
+jest.mock('src/classify');
+jest.mock('../paymentMethodSummary', () => 'PaymentMethodSummary');
+jest.mock('../shippingAddressSummary', () => 'ShippingAddressSummary');
+jest.mock('../shippingMethodSummary', () => 'shippingMethodSummary');
+
+const mockCancelCheckout = jest.fn();
+const mockSetEditing = jest.fn();
+const mockSubmitOrder = jest.fn();
+
+const defaultProps = {
+    cancelCheckout: mockCancelCheckout,
+    cart: {
+        details: {},
+        cartId: '123',
+        totals: {
+            subtotal: 3.5
+        }
+    },
+    classes: {
+        body: 'body',
+        footer: 'footer'
+    },
+    setEditing: mockSetEditing,
+    submitOrder: mockSubmitOrder,
+    submitting: false
+};
+
+beforeEach(() => {
+    mockCancelCheckout.mockReset();
+    mockSetEditing.mockReset();
+    mockSubmitOrder.mockReset();
+});
+
+test('renders an Overview component', () => {
+    const component = createTestInstance(<Overview {...defaultProps} />);
+
+    expect(component.toJSON()).toMatchSnapshot();
+});
+
+test('renders "0" for Price if subtotal is falsy', () => {
+    const props = {
+        ...defaultProps,
+        cart: {
+            totals: {},
+            details: {}
+        }
+    };
+    const component = createTestInstance(<Overview {...props} />);
+
+    expect(component.toJSON()).toMatchSnapshot();
+});
+
+test('Confirm Order button is disabled if submitting', () => {
+    const props = {
+        ...defaultProps,
+        submitting: true
+    };
+    const component = createTestInstance(<Overview {...props} />);
+    expect(component.toJSON()).toMatchSnapshot();
+});
+
+test('Confirm Order button is disabled if not ready', () => {
+    const props = {
+        ...defaultProps,
+        ready: false
+    };
+    const component = createTestInstance(<Overview {...props} />);
+    expect(component.toJSON()).toMatchSnapshot();
+});
+
+test('clicking address form edit sets `editing` state value to "address"', () => {
+    const component = createTestInstance(<Overview {...defaultProps} />);
+
+    act(() => {
+        component.root.findAllByType(Section)[0].props.onClick();
+    });
+
+    expect(mockSetEditing).toHaveBeenCalledWith('address');
+});
+
+test('clicking payment form edit sets `editing` state value to "paymentMethod"', () => {
+    const component = createTestInstance(<Overview {...defaultProps} />);
+
+    act(() => {
+        component.root.findAllByType(Section)[1].props.onClick();
+    });
+
+    expect(mockSetEditing).toHaveBeenCalledWith('paymentMethod');
+});
+
+test('clicking shipping form edit sets `editing` state value to "shippingMethod"', () => {
+    const component = createTestInstance(<Overview {...defaultProps} />);
+
+    act(() => {
+        component.root.findAllByType(Section)[2].props.onClick();
+    });
+
+    expect(mockSetEditing).toHaveBeenCalledWith('shippingMethod');
+});
